error 403 fixed
Fixed error 403 due to axios and CORS
Used fetch instead of axios to fix the issue
